feat: combined image preview and dropzone input
