+++ components/common/index.ts
@@ -8,3 +8,4 @@ export { default as Switch } from "./Switch";
 export { default as Overlay } from "./Overlay";
 export { default as NavLink } from "./NavLink";
 export { default as ProgressBar } from "./ProgressBar";
+export * from "./DropZoneInput";
