+++ components/ui/index.ts
@@ -1,9 +1,9 @@
 export * as Accordion from "./accordion";
-export { default as Card } from "./card";
 export { default as Button } from "./button";
-export * as Form from "./form";
-export * as Select from "./select";
-export * as DropdownMenu from "./dropdown-menu";
-export * as Popover from "./popover";
+export { default as Card } from "./card";
 export { DatePicker } from "./date-picker";
-export { default as DropZone } from "./drop-zone";
+export { DropZone } from "./drop-zone";
+export * as DropdownMenu from "./dropdown-menu";
+export * as Form from "./form";
+export * as Popover from "./popover";
+export * as Select from "./select";
